ref: rename middlewares to middleware
