+++ src/app/routes/public-forms.server.routes.js
@@ -17,7 +17,7 @@ const { rateLimitConfig } = require('../../config/config')
 const PublicFormController = require('../modules/form/public-form/public-form.controller')
 const SpcpController = require('../modules/spcp/spcp.controller')
 const { BasicField } = require('../../types')
-const EmailSubmissionsMiddleware = require('../../app/modules/submission/email-submission/email-submission.middlewares')
+const EmailSubmissionsMiddleware = require('../../app/modules/submission/email-submission/email-submission.middleware')
 
 module.exports = function (app) {
   /**
