+++ tests/unit/backend/controllers/email-submissions.server.controller.spec.js
@@ -10,7 +10,7 @@ const dbHandler = require('../helpers/db-handler')
 const { ObjectID } = require('bson-ext')
 const MailService = require('../../../../dist/backend/app/services/mail/mail.service')
   .default
-const EmailSubmissionsMiddleware = require('../../../../dist/backend/app/modules/submission/email-submission/email-submission.middlewares')
+const EmailSubmissionsMiddleware = require('../../../../dist/backend/app/modules/submission/email-submission/email-submission.middleware')
 const User = dbHandler.makeModel('user.server.model', 'User')
 const Agency = dbHandler.makeModel('agency.server.model', 'Agency')
 const Form = dbHandler.makeModel('form.server.model', 'Form')
